v9.385: the widening the schema alone could not tell me about
v9.384 widened five surrogate ids and broke nine CI jobs, all for one reason:
every job loads the schema, and uc7_warrant_audit declares RETURNS TABLE
(event_id INTEGER, ...) while now receiving a BIGINT. Postgres refuses with
"structure of query does not match function result type" the first time the
procedure file is loaded.

Widening a column touches three things the column definition does not show.

FIVE ROW-RETURNING FUNCTIONS declare these ids in their result columns, and
CREATE OR REPLACE cannot change a function's return type. The migration now
drops them and the object sync recreates them -- and they are FOUND rather than
listed. A TABLE-returning function's result columns are OUT arguments in the
catalog, so the ones exposing a widened id as 32-bit can be asked for by name and
type. The first draft listed five signatures by hand and four were wrong, which
is worse than useless: DROP FUNCTION IF EXISTS with a signature that does not
match says "does not exist, skipping" and leaves the broken function in place.

THREE VIEWS depend on these columns and Postgres refuses ALTER COLUMN TYPE while
they do. The migration captures each view's definition AND ITS GRANTS from the
live catalog, drops it, alters, recreates. From the catalog rather than a copy,
because a copy is right the day it is written and wrong the first time somebody
edits the view. polaris_app holds real privileges on all three.

THE COLUMN WORK IS NOW CONDITIONAL on a target still being integer, so applying
this to a database loaded from the current schema touches no view at all. The
sequence widening stays unconditional: a sequence left at the 32-bit ceiling
under a 64-bit column is the half-finished state the migration exists to avoid.

And the migration says it is NOT zero-downtime rather than implying otherwise.
Between --up and --sync-objects those functions do not exist.

The underlying failure was mine: I verified the widening against the schema file
and the model and never loaded it into a database, because no local interpreter
has Flask. Loading the SQL needs only psql, which was available the whole time.
Both paths are now exercised locally before pushing: a fresh load, and an upgrade
from the pre-widening schema through the migration and the object sync, ending
with a real row inserted at event_id 2,147,483,648.

+++ polaris_sql/05_procedures.sql
@@ -424,7 +424,7 @@ CREATE OR REPLACE FUNCTION uc7_warrant_audit(
     p_window_end         TIMESTAMP DEFAULT 'infinity',
     p_context_filter     VARCHAR(40) DEFAULT NULL  -- if non-null, restrict to one context
 ) RETURNS TABLE (
-    event_id           INTEGER,
+    event_id          BIGINT,
     event_timestamp    TIMESTAMP,
     context_type       VARCHAR(40),
     requesting_agency  VARCHAR(200),
